Move new editor variants inside /ts/editor directory
