+++ qt/aqt/data/web/css/BUILD.bazel
@@ -24,27 +24,13 @@ copy_files_into_group(
 )
 
 copy_files_into_group(
-    name = "browser_editor",
+    name = "editor",
     srcs = [
         "browser_editor.css",
-    ],
-    package = "//ts/browser-editor",
-)
-
-copy_files_into_group(
-    name = "reviewer_editor",
-    srcs = [
         "reviewer_editor.css",
-    ],
-    package = "//ts/reviewer-editor",
-)
-
-copy_files_into_group(
-    name = "note_creator",
-    srcs = [
         "note_creator.css",
     ],
-    package = "//ts/note-creator",
+    package = "//ts/editor",
 )
 
 copy_files_into_group(
@@ -60,10 +46,8 @@ filegroup(
     srcs = [
         "css_local",
         "editable",
+        "editor",
         "reviewer",
-        "browser_editor",
-        "reviewer_editor",
-        "note_creator",
     ],
     visibility = ["//qt:__subpackages__"],
 )
