--- ts/browser-editor/BUILD.bazel
@@ -1,69 +1,0 @@
-load("//ts/svelte:svelte.bzl", "compile_svelte", "svelte_check")
-load("//ts:prettier.bzl", "prettier_test")
-load("//ts:eslint.bzl", "eslint_test")
-load("//ts:esbuild.bzl", "esbuild")
-load("//ts:typescript.bzl", "typescript")
-
-_ts_deps = [
-    "//ts/editor",
-    "//ts/components",
-    "//ts/lib",
-    "//ts/domlib",
-    "//ts/sveltelib",
-    "@npm//@fluent",
-    "@npm//codemirror",
-    "@npm//@types/codemirror",
-    "@npm//svelte",
-]
-
-compile_svelte(deps = _ts_deps)
-
-typescript(
-    name = "browser_editor_ts",
-    deps = _ts_deps + [
-        ":svelte",
-    ],
-)
-
-esbuild(
-    name = "browser_editor",
-    args = {
-        "loader": {".svg": "text"},
-    },
-    entry_point = "index_wrapper.ts",
-    output_css = "browser_editor.css",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":browser_editor_ts",
-        "//ts/editor",
-        "//ts/editor:editor_css",
-        "//sass:button_mixins_lib",
-        "@npm//@mdi",
-        "@npm//bootstrap-icons",
-        "@npm//protobufjs",
-    ],
-)
-
-# Tests
-################
-
-prettier_test()
-
-eslint_test()
-
-svelte_check(
-    name = "svelte_check",
-    srcs = glob([
-        "**/*.ts",
-        "**/*.svelte",
-    ]) + [
-        "//sass:button_mixins_lib",
-        "//sass/bootstrap",
-        "//ts/components",
-        "//ts/editor",
-        "//ts/editable:editable_ts",
-        "@npm//@types/bootstrap",
-        "@npm//@types/codemirror",
-        "@npm//codemirror",
-    ],
-)
