+++ ts/editor/BUILD.bazel
@@ -41,6 +41,48 @@ typescript(
     ],
 )
 
+_esbuild_deps = [
+    ":editor",
+    ":editor_css",
+    "//sass:button_mixins_lib",
+    "@npm//@mdi",
+    "@npm//bootstrap-icons",
+    "@npm//protobufjs",
+]
+
+esbuild(
+    name = "browser_editor",
+    args = {
+        "loader": {".svg": "text"},
+    },
+    entry_point = "index_browser.ts",
+    output_css = "browser_editor.css",
+    visibility = ["//visibility:public"],
+    deps = _esbuild_deps,
+)
+
+esbuild(
+    name = "reviewer_editor",
+    args = {
+        "loader": {".svg": "text"},
+    },
+    entry_point = "index_reviewer.ts",
+    output_css = "reviewer_editor.css",
+    visibility = ["//visibility:public"],
+    deps = _esbuild_deps,
+)
+
+esbuild(
+    name = "note_creator",
+    args = {
+        "loader": {".svg": "text"},
+    },
+    entry_point = "index_creator.ts",
+    output_css = "note_creator.css",
+    visibility = ["//visibility:public"],
+    deps = _esbuild_deps,
+)
+
 # Tests
 ################
