+++ ts/editor/BrowserEditor.svelte
