+++ ts/editor/NoteCreator.svelte
