--- ts/note-creator/BUILD.bazel
@@ -1,69 +1,0 @@
-load("//ts/svelte:svelte.bzl", "compile_svelte", "svelte_check")
-load("//ts:prettier.bzl", "prettier_test")
-load("//ts:eslint.bzl", "eslint_test")
-load("//ts:esbuild.bzl", "esbuild")
-load("//ts:typescript.bzl", "typescript")
-
-_ts_deps = [
-    "//ts/editor",
-    "//ts/components",
-    "//ts/lib",
-    "//ts/domlib",
-    "//ts/sveltelib",
-    "@npm//@fluent",
-    "@npm//codemirror",
-    "@npm//@types/codemirror",
-    "@npm//svelte",
-]
-
-compile_svelte(deps = _ts_deps)
-
-typescript(
-    name = "note_creator_ts",
-    deps = _ts_deps + [
-        ":svelte",
-    ],
-)
-
-esbuild(
-    name = "note_creator",
-    args = {
-        "loader": {".svg": "text"},
-    },
-    entry_point = "index_wrapper.ts",
-    output_css = "note_creator.css",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":note_creator_ts",
-        "//ts/editor",
-        "//ts/editor:editor_css",
-        "//sass:button_mixins_lib",
-        "@npm//@mdi",
-        "@npm//bootstrap-icons",
-        "@npm//protobufjs",
-    ],
-)
-
-# Tests
-################
-
-prettier_test()
-
-eslint_test()
-
-svelte_check(
-    name = "svelte_check",
-    srcs = glob([
-        "**/*.ts",
-        "**/*.svelte",
-    ]) + [
-        "//sass:button_mixins_lib",
-        "//sass/bootstrap",
-        "//ts/components",
-        "//ts/editor",
-        "//ts/editable:editable_ts",
-        "@npm//@types/bootstrap",
-        "@npm//@types/codemirror",
-        "@npm//codemirror",
-    ],
-)
