--- ts/reviewer-editor/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "include": ["*"],
-    "references": [
-        { "path": "../lib" },
-        { "path": "../domlib" },
-        { "path": "../sveltelib" },
-        { "path": "../components" },
-        { "path": "../editor" }
-    ]
-}
